fix(test): run tests and change structure of spec file (#4663)
