+++ package.json
@@ -40,7 +40,7 @@
     "git-commit-changelog": "git commit -am \"chore(changelog): update [skip ci] \"",
     "changelog": "conventional-changelog -i CHANGELOG.md -s -p angular -r 2",
     "cy:open": "cypress open",
-    "cy:run": "cypress run --config videoRecording=false",
+    "cy:run": "cypress run --config video=false",
     "view-stats": "webpack-bundle-analyzer demo/dist/stats.json",
     "build:dynamic": "run-s build link build:client-and-server-bundles webpack:server configure-heroku",
     "serve:dynamic": "node demo/dist/server.js",
@@ -112,7 +112,7 @@
     "core-js": "2.5.6",
     "cpy": "7.0.0",
     "cpy-cli": "1.0.1",
-    "cypress": "2.1.0",
+    "cypress": "3.1.0",
     "gh-pages": "1.1.0",
     "gitignore-to-glob": "0.3.0",
     "google-code-prettify": "1.0.5",
